In works animation of circles rotating within circles.
Designed to be an animated score for musical interpretation.

Currently has lots of issues; but generates gray 'planets' that have orbiting 'satellites'.

In the long run, will implement a stochastic engine to turn on and off the satellites for various amounts of time. Also, I intend to feature creep a GUI.